requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
requests==2.21.0
numpy==1.17.0
gcloud==0.18.3
firebase==3.0.1
googleapis-common-protos==1.6.0
sseclient==0.0.24
python-jwt
six==1.9.0
requests-toolbelt==0.9.1
firebase_admin
PyYaml==5.1.2
